break graphify out of stbernard
